% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_reddit_data.R
\name{clean_reddit_data}
\alias{clean_reddit_data}
\title{clean_reddit_data}
\usage{
clean_reddit_data(reddit_data = NULL)
}
\arguments{
\item{reddit_data}{A dataframe of reddit data}
}
\value{
data.frame
}
\description{
clean_reddit_data
}
