% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_crawl_submissions.R
\name{reddit_crawl_submissions}
\alias{reddit_crawl_submissions}
\title{reddit_crawl_submissions}
\usage{
reddit_crawl_submissions(reddit_con, subreddit = NULL, limit = 10)
}
\arguments{
\item{reddit_con}{Connection to Reddit}

\item{subreddit}{A subreddit name, defaults to all}

\item{limit}{Amount of submissions you want}
}
\value{
data.frame
}
\description{
reddit_crawl_submissions
}
